feat: extract the OtpParser.kt from OtpProcessor.kt for better management; removed CLIPBOARD_OTP_CODE_KEY that was redundant.
